+++ utoipa-gen/src/lib.rs
@@ -53,6 +53,7 @@ mod openapi;
 mod path;
 mod schema_type;
 mod security_requirement;
+mod server;
 
 use crate::path::{Path, PathAttr};
 
@@ -69,7 +70,7 @@ use self::{
 static CONFIG: once_cell::sync::Lazy<utoipa_config::Config> =
     once_cell::sync::Lazy::new(utoipa_config::Config::read_from_file);
 
-#[proc_macro_derive(ToSchema, attributes(schema))]
+#[proc_macro_derive(ToSchema, attributes(schema, serde))]
 /// Generate reusable OpenAPI schema to be used
 /// together with [`OpenApi`][openapi_derive].
 ///
@@ -135,7 +136,7 @@ static CONFIG: once_cell::sync::Lazy<utoipa_config::Config> =
 /// * `min_properties = ...` Can be used to define minimum number of properties this struct can
 ///   contain. Value must be a number.
 ///* `no_recursion` Is used to break from recursion in case of looping schema tree e.g. `Pet` ->
-///  `Owner` -> `Pet`. _`no_recursion`_ attribute must be used within `Ower` type not to allow
+///  `Owner` -> `Pet`. _`no_recursion`_ attribute must be used within `Owner` type not to allow
 ///  recurring into `Pet`. Failing to do so will cause infinite loop and runtime **panic**. On
 ///  struct level the _`no_recursion`_ rule will be applied to all of its fields.
 ///
@@ -199,10 +200,10 @@ static CONFIG: once_cell::sync::Lazy<utoipa_config::Config> =
 ///   See [`Object::content_encoding`][schema_object_encoding]
 /// * `content_media_type = ...` Can be used to define MIME type of a string for underlying schema object.
 ///   See [`Object::content_media_type`][schema_object_media_type]
-///* `ignore` or `ignore = ...` Can be used to skip the field from being serialized to OpenAPI schema. It accepts either a literal `bool` value
-///   or a path to a function that returns `bool` (`Fn() -> bool`).
-///* `no_recursion` Is used to break from recursion in case of looping schema tree e.g. `Pet` ->
-///  `Owner` -> `Pet`. _`no_recursion`_ attribute must be used within `Ower` type not to allow
+/// * `ignore` or `ignore = ...` Can be used to skip the field from being serialized to OpenAPI schema. (Currently it accepts either a literal `bool` value
+///   or a path to a function that returns `bool` (`Fn() -> bool`). **Note!** support for function paths is **deprecated** and will be removed in a future version.).
+/// * `no_recursion` Is used to break from recursion in case of looping schema tree e.g. `Pet` ->
+///  `Owner` -> `Pet`. _`no_recursion`_ attribute must be used within `Owner` type not to allow
 ///  recurring into `Pet`. Failing to do so will cause infinite loop and runtime **panic**.
 ///
 /// #### Field nullability and required rules
@@ -267,7 +268,7 @@ static CONFIG: once_cell::sync::Lazy<utoipa_config::Config> =
 /// * `content_media_type = ...` Can be used to define MIME type of a string for underlying schema object.
 ///   See [`Object::content_media_type`][schema_object_media_type]
 ///* `no_recursion` Is used to break from recursion in case of looping schema tree e.g. `Pet` ->
-///  `Owner` -> `Pet`. _`no_recursion`_ attribute must be used within `Ower` type not to allow
+///  `Owner` -> `Pet`. _`no_recursion`_ attribute must be used within `Owner` type not to allow
 ///  recurring into `Pet`. Failing to do so will cause infinite loop and runtime **panic**.
 ///
 /// # Enum Optional Configuration Options for `#[schema(...)]`
@@ -335,7 +336,7 @@ static CONFIG: once_cell::sync::Lazy<utoipa_config::Config> =
 ///   field for enums with single unnamed _`ToSchema`_ reference field. See the [discriminator
 ///   syntax][derive@ToSchema#schemadiscriminator-syntax].
 ///* `no_recursion` Is used to break from recursion in case of looping schema tree e.g. `Pet` ->
-///  `Owner` -> `Pet`. _`no_recursion`_ attribute must be used within `Ower` type not to allow
+///  `Owner` -> `Pet`. _`no_recursion`_ attribute must be used within `Owner` type not to allow
 ///  recurring into `Pet`. Failing to do so will cause infinite loop and runtime **panic**. On
 ///  enum level the _`no_recursion`_ rule will be applied to all of its variants.
 ///
@@ -391,7 +392,7 @@ static CONFIG: once_cell::sync::Lazy<utoipa_config::Config> =
 /// * `min_properties = ...` Can be used to define minimum number of properties this struct can
 ///   contain. Value must be a number.
 ///* `no_recursion` Is used to break from recursion in case of looping schema tree e.g. `Pet` ->
-///  `Owner` -> `Pet`. _`no_recursion`_ attribute must be used within `Ower` type not to allow
+///  `Owner` -> `Pet`. _`no_recursion`_ attribute must be used within `Owner` type not to allow
 ///  recurring into `Pet`. Failing to do so will cause infinite loop and runtime **panic**. On
 ///  named field variant level the _`no_recursion`_ rule will be applied to all of its fields.
 ///
@@ -422,7 +423,7 @@ static CONFIG: once_cell::sync::Lazy<utoipa_config::Config> =
 ///   not in the code. If you'd like to mark the field as deprecated in the code as well use
 ///   Rust's own `#[deprecated]` attribute instead.
 ///* `no_recursion` Is used to break from recursion in case of looping schema tree e.g. `Pet` ->
-///  `Owner` -> `Pet`. _`no_recursion`_ attribute must be used within `Ower` type not to allow
+///  `Owner` -> `Pet`. _`no_recursion`_ attribute must be used within `Owner` type not to allow
 ///  recurring into `Pet`. Failing to do so will cause infinite loop and runtime **panic**.
 ///
 /// #### Mixed Enum Unnamed Field Variant's Field Configuration Options
@@ -1038,6 +1039,8 @@ pub fn derive_to_schema(input: TokenStream) -> TokenStream {
 /// * `description = ...` Allows overriding description of the path. Value can be literal string or valid
 ///   rust expression e.g. `include_str!(...)` or `const` reference.
 ///
+/// * `extensions(...)` List of extensions local to the path operation.
+///
 /// # Request Body Attributes
 ///
 /// ## Simple format definition by `request_body = ...`
@@ -1350,6 +1353,8 @@ pub fn derive_to_schema(input: TokenStream) -> TokenStream {
 /// * `example = ...` Can method reference or _`json!(...)`_. Given example
 ///   will override any example in underlying parameter type.
 ///
+/// * `extensions(...)` List of extensions local to the parameter
+///
 /// ##### Parameter type attributes
 ///
 /// These attributes supported when _`parameter_type`_ is present. Either by manually providing one
@@ -1469,6 +1474,19 @@ pub fn derive_to_schema(input: TokenStream) -> TokenStream {
 /// (),
 /// ("api_key" = []),
 /// ("key" = [], "key2" = []),
+/// ```
+///
+/// # Extensions Requirements Attributes
+///
+/// * `x-property` defines the name of the extension.
+/// * `json!(...)` defines the value associated with the named extension as a `serde_json::Value`.
+///
+/// **Extensions Requitement supported formats:**
+///
+/// ```text
+/// ("x-property" = json!({ "type": "mock" }) ),
+/// ("x-an-extension" = json!({ "type": "mock" }) ),
+/// ("x-another-extension" = json!( "body" ) ),
 /// ```
 ///
 /// # actix_extras feature support for actix-web
@@ -2364,8 +2382,8 @@ pub fn openapi(input: TokenStream) -> TokenStream {
 ///   Free form type enables use of arbitrary types within map values.
 ///   Supports formats _`additional_properties`_ and _`additional_properties = true`_.
 ///
-/// * `ignore` or `ignore = ...` Can be used to skip the field from being serialized to OpenAPI schema. It accepts either a literal `bool` value
-///   or a path to a function that returns `bool` (`Fn() -> bool`).
+/// * `ignore` or `ignore = ...` Can be used to skip the field from being serialized to OpenAPI schema. (Currently it accepts either a literal `bool` value
+///   or a path to a function that returns `bool` (`Fn() -> bool`). **Note!** support for function paths is **deprecated** and will be removed in a future version.).
 ///
 /// #### Field nullability and required rules
 ///
@@ -3754,7 +3772,7 @@ mod parse_utils {
         Punctuated::parse_terminated(&group)
     }
 
-    pub fn parse_comma_separated_within_parethesis_with<T>(
+    pub fn parse_comma_separated_within_parenthesis_with<T>(
         input: ParseStream,
         with: fn(ParseStream) -> syn::Result<T>,
     ) -> syn::Result<Punctuated<T, Comma>>