+++ utoipa-swagger-ui/build.rs
@@ -2,7 +2,7 @@ use std::{
     env,
     error::Error,
     fs::{self, File},
-    io::{self, Cursor},
+    io::{self, Cursor, Read, Seek},
     path::{Path, PathBuf},
 };
 
@@ -81,71 +81,68 @@ enum SwaggerZip {
 }
 
 impl SwaggerZip {
-    fn len(&self) -> usize {
-        match self {
-            Self::File(file) => file.len(),
-            Self::Bytes(bytes) => bytes.len(),
-        }
-    }
-
-    fn by_index(&mut self, index: usize) -> Result<zip::read::ZipFile, ZipError> {
-        match self {
-            Self::File(file) => file.by_index(index),
-            Self::Bytes(bytes) => bytes.by_index(index),
-        }
-    }
-
     fn extract_dist(&mut self, target_dir: &str) -> Result<String, ZipError> {
-        let mut zip_top_level_folder = String::new();
+        // Inner function that's generic over the type of zip files
+        fn extract<R: Seek + Read>(
+            zip: &mut ZipArchive<R>,
+            target_dir: &str,
+        ) -> Result<String, ZipError> {
+            let mut zip_top_level_folder = String::new();
 
-        for index in 0..self.len() {
-            let mut file = self.by_index(index)?;
-            let filepath = file
-                .enclosed_name()
-                .ok_or(ZipError::InvalidArchive("invalid path file".into()))?;
+            for index in 0..zip.len() {
+                let mut file = zip.by_index(index)?;
+                let filepath = file
+                    .enclosed_name()
+                    .ok_or(ZipError::InvalidArchive("invalid path file".into()))?;
 
-            if index == 0 {
-                zip_top_level_folder = filepath
+                if index == 0 {
+                    zip_top_level_folder = filepath
+                        .iter()
+                        .take(1)
+                        .map(|x| x.to_str().unwrap_or_default())
+                        .collect::<String>();
+                }
+
+                let next_folder = filepath
                     .iter()
+                    .skip(1)
                     .take(1)
                     .map(|x| x.to_str().unwrap_or_default())
                     .collect::<String>();
-            }
 
-            let next_folder = filepath
-                .iter()
-                .skip(1)
-                .take(1)
-                .map(|x| x.to_str().unwrap_or_default())
-                .collect::<String>();
+                if next_folder == "dist" {
+                    let directory = [&target_dir].iter().collect::<PathBuf>();
+                    let out_path = directory.join(filepath);
 
-            if next_folder == "dist" {
-                let directory = [&target_dir].iter().collect::<PathBuf>();
-                let out_path = directory.join(filepath);
-
-                if file.name().ends_with('/') {
-                    fs::create_dir_all(&out_path)?;
-                } else {
-                    if let Some(p) = out_path.parent() {
-                        if !p.exists() {
-                            fs::create_dir_all(p)?;
+                    if file.name().ends_with('/') {
+                        fs::create_dir_all(&out_path)?;
+                    } else {
+                        if let Some(p) = out_path.parent() {
+                            if !p.exists() {
+                                fs::create_dir_all(p)?;
+                            }
+                        }
+                        let mut out_file = fs::File::create(&out_path)?;
+                        io::copy(&mut file, &mut out_file)?;
+                    }
+                    // Get and Set permissions
+                    #[cfg(unix)]
+                    {
+                        use std::os::unix::fs::PermissionsExt;
+                        if let Some(mode) = file.unix_mode() {
+                            fs::set_permissions(&out_path, fs::Permissions::from_mode(mode))?;
                         }
                     }
-                    let mut out_file = fs::File::create(&out_path)?;
-                    io::copy(&mut file, &mut out_file)?;
-                }
-                // Get and Set permissions
-                #[cfg(unix)]
-                {
-                    use std::os::unix::fs::PermissionsExt;
-                    if let Some(mode) = file.unix_mode() {
-                        fs::set_permissions(&out_path, fs::Permissions::from_mode(mode))?;
-                    }
                 }
             }
+
+            Ok(zip_top_level_folder)
         }
 
-        Ok(zip_top_level_folder)
+        match self {
+            Self::File(file) => extract(file, target_dir),
+            Self::Bytes(bytes) => extract(bytes, target_dir),
+        }
     }
 }