+++ utoipa-swagger-ui/CHANGELOG.md
@@ -1,5 +1,22 @@
 # Changelog - utoipa-swagger-ui
 
+## 9.0.2 - Sun 25 May 2025
+
+### Fixed
+
+* Avoid cloning ApiDoc on every request for axum (https://github.com/juhaku/utoipa/pull/1393)
+
+### Changed
+
+* Update build dependency `zip` to v3.0 (https://github.com/juhaku/utoipa/pull/1391)
+
+## 9.0.1 - Fri 4 Apr 2025
+
+### Fixed
+
+* fix(swagger-ui): fix build for breaking zip crate 2.6.0 change (#1363)
+* fix: mismatch error from InvalidArchive (#1343)
+
 ## 9.0.0 - Thu 16 2025
 
 ### Changed